requirements.txt
# Core Image Processing
opencv-python==4.9.0.80 
numpy==1.26.4
pillow==10.0.1
matplotlib==3.8.2

# YOLOv8 Object Detection
ultralytics==8.0.237

# PyTorch (YOLOv8 uses this)
torch==2.1.1
torchvision==0.16.1

# Configuration & Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm

# API (optional)
fastapi
uvicorn[standard]

# Development (optional)
jupyter==1.0.0
scikit-learn
albumentations